+++ CHANGELOG.md
@@ -9,10 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `lists:keysort/2`
+- Added `lists:merge/2,3`
 
 ### Fixed
 
 - Fixed a bug where binary matching could fail due to a missing preservation of the matched binary.
+
+### Changed
+
+- lists sort function now use a stable merge sort implementation instead of quick sort
 
 ## [0.6.6] - 2025-06-23